+++ src/api/readme.md
@@ -1,0 +1,1 @@
+# 해당 폴더에는 api 통신에 대한 코드를 작성해주세요
