--- src/components/table/QA.tsx
@@ -1,9 +1,0 @@
-type QAType = {
-  shopCode: string;
-};
-
-const QA = ({ shopCode }: QAType) => {
-  return <div>{shopCode}</div>;
-};
-
-export default QA;
