+++ src/features/mainPage/ShopAction.tsx
@@ -1,0 +1,9 @@
+interface ShopActionType {
+  shopCode: string;
+}
+
+const ShopAction = ({ shopCode }: ShopActionType) => {
+  return <div>{shopCode}</div>;
+};
+
+export default ShopAction;
