+++ src/features/mainPage/styles/Styles.ts
@@ -1,10 +1,5 @@
 import styled from 'styled-components';
 
-export const Search = styled.section`
-  display: flex;
-  justify-content: center;
-  padding-top: 20px;
-`;
 export const ContentsArea = styled.section`
   height: 70vh;
   overflow: auto;
