+++ src/layout/mainLayout/styles/MainLayoutStyle.tsx
@@ -1,0 +1,9 @@
+import { Layout } from 'antd';
+import styled from 'styled-components';
+
+export const PageContainer = styled.div`
+  margin: 0 auto;
+  padding: 40px;
+  width: 90%;
+  min-width: 1024px;
+`;
