+++ src/pages/mainPage/Mainpage.tsx
@@ -1,27 +1,13 @@
-import ShopList from 'components/table/ShopList';
-import * as S from './MainPageStyle';
-import RadioFilter from 'components/filter/RadioFilter';
+import Container from 'features/mainPage/Container';
 
 const MainPage = () => {
   return (
     <main>
       <h1>QASys</h1>
       <hr />
-      <S.ContentsArea>
-        <S.FilterArea>
-          <RadioFilter options={FILTER_LIST} />
-        </S.FilterArea>
-        <ShopList />
-      </S.ContentsArea>
+      <Container />
     </main>
   );
 };
 
 export default MainPage;
-
-const FILTER_LIST = [
-  { label: '10', value: 10 },
-  { label: '20', value: 20 },
-  { label: '50', value: 50 },
-  { label: '100', value: 100 },
-];
