--- src/pages/mainPage/interface.ts
@@ -1,5 +1,0 @@
-type Menu = { label: string; value: string };
-
-export interface SearchMenu {
-  searchMenu: Menu[];
-}
