+++ src/utils/const/MainPage.ts
@@ -1,0 +1,6 @@
+export const FILTER_LIST = [
+  { label: '10', value: 10 },
+  { label: '20', value: 20 },
+  { label: '50', value: 50 },
+  { label: '100', value: 100 },
+];
