+++ src/utils/mock/MockData.ts
@@ -1,0 +1,92 @@
+import type { DataType } from 'features/mainPage/types/types';
+
+export const shopListData: DataType[] = [
+  {
+    key: 1,
+    shopName: '옥션',
+    shopCode: 'A001',
+    isApi: 'X',
+    isAppliedEngine: 'X',
+    qa: '',
+  },
+  {
+    key: 2,
+    shopName: 'CJMall',
+    shopCode: 'A004',
+    isApi: 'X',
+    isAppliedEngine: 'O',
+    qa: '',
+  },
+  {
+    key: 3,
+    shopName: 'GMarket',
+    shopCode: 'A006',
+    isApi: 'X',
+    isAppliedEngine: 'X',
+    qa: '',
+  },
+  {
+    key: 4,
+    shopName: '스마트스토어',
+    shopCode: 'A077',
+    isApi: 'O',
+    isAppliedEngine: 'X',
+    qa: '',
+  },
+  {
+    key: 5,
+    shopName: '11번가',
+    shopCode: 'A112',
+    isApi: 'O',
+    isAppliedEngine: 'O',
+    qa: '',
+  },
+  {
+    key: 6,
+    shopName: '11번가(신규)',
+    shopCode: 'A113',
+    isApi: 'O',
+    isAppliedEngine: 'X',
+    qa: '',
+  },
+  {
+    key: 7,
+    shopName: '현대H몰',
+    shopCode: 'A131',
+    isApi: 'O',
+    isAppliedEngine: 'O',
+    qa: '',
+  },
+  {
+    key: 8,
+    shopName: '위메프2.0',
+    shopCode: 'B719',
+    isApi: 'O',
+    isAppliedEngine: 'X',
+    qa: '',
+  },
+  {
+    key: 9,
+    shopName: '카카오톡스토어',
+    shopCode: 'B688',
+    isApi: 'O',
+    isAppliedEngine: 'X',
+    qa: '',
+  },
+  {
+    key: 10,
+    shopName: '티몬',
+    shopCode: 'B956',
+    isApi: 'O',
+    isAppliedEngine: 'X',
+    qa: '',
+  },
+  {
+    key: 11,
+    shopName: '오늘의집',
+    shopCode: 'B959',
+    isApi: 'O',
+    isAppliedEngine: 'O',
+    qa: '',
+  },
+];
